Refactor demo into per-tab components
Each tab (PinchPan, Carousel, ScalableCarousel) now lives in its own
file. App.tsx is reduced to tab state and nav.
